fixup! Only work with fully expanded PCollections in TransformHierarchy and in PTransformReplacements
