+++ sdks/java/core/src/main/java/org/apache/beam/sdk/values/PCollectionViews.java
@@ -41,9 +41,7 @@ import java.util.PriorityQueue;
 import java.util.RandomAccess;
 import java.util.Set;
 import java.util.SortedMap;
-import java.util.concurrent.atomic.AtomicReference;
 import java.util.function.Supplier;
-import org.apache.beam.sdk.Pipeline;
 import org.apache.beam.sdk.annotations.Experimental;
 import org.apache.beam.sdk.annotations.Experimental.Kind;
 import org.apache.beam.sdk.annotations.Internal;
@@ -52,8 +50,6 @@ import org.apache.beam.sdk.coders.Coder;
 import org.apache.beam.sdk.coders.CoderException;
 import org.apache.beam.sdk.coders.StructuredCoder;
 import org.apache.beam.sdk.io.range.OffsetRange;
-import org.apache.beam.sdk.runners.AppliedPTransform;
-import org.apache.beam.sdk.runners.TransformHierarchy;
 import org.apache.beam.sdk.transforms.Materialization;
 import org.apache.beam.sdk.transforms.Materializations;
 import org.apache.beam.sdk.transforms.Materializations.IterableView;
@@ -1426,56 +1422,5 @@ public class PCollectionViews {
         }
       };
     }
-  }
-
-  public static <InputT, ViewT> PCollectionView<ViewT> findPCollectionView(
-      final AppliedPTransform<
-              PCollection<InputT>,
-              PCollectionView<ViewT>,
-              ? extends PTransform<PCollection<InputT>, PCollectionView<ViewT>>>
-          transform) {
-    final AtomicReference<PCollectionView<ViewT>> viewRef = new AtomicReference<>();
-    transform
-        .getPipeline()
-        .traverseTopologically(
-            new Pipeline.PipelineVisitor.Defaults() {
-              // Stores whether we have entered the expected composite view transform.
-              private boolean tracking = false;
-
-              @Override
-              public CompositeBehavior enterCompositeTransform(TransformHierarchy.Node node) {
-                if (transform.getTransform() == node.getTransform()) {
-                  tracking = true;
-                }
-                return super.enterCompositeTransform(node);
-              }
-
-              @Override
-              public void visitPrimitiveTransform(TransformHierarchy.Node node) {
-                if (tracking && node.getTransform() instanceof View.CreatePCollectionView) {
-                  View.CreatePCollectionView createViewTransform =
-                      (View.CreatePCollectionView) node.getTransform();
-                  checkState(
-                      viewRef.compareAndSet(null, createViewTransform.getView()),
-                      "Found more than one instance of a CreatePCollectionView when"
-                          + "attempting to replace %s, found [%s, %s]",
-                      transform.getTransform(),
-                      viewRef.get(),
-                      createViewTransform.getView());
-                }
-              }
-
-              @Override
-              public void leaveCompositeTransform(TransformHierarchy.Node node) {
-                if (transform.getTransform() == node.getTransform()) {
-                  tracking = false;
-                }
-              }
-            });
-    checkState(
-        viewRef.get() != null,
-        "Expected to find CreatePCollectionView contained within %s",
-        transform.getTransform());
-    return viewRef.get();
   }
 }
